Fixing issue with book not starting from the right time
When using the Siri intent AND the app is not in the background when the app starts it plays from the beginning of the previous book but not form where you played last.

This creates a better flag to check if the book should start playing or if it should postpone.